Added an internal error handler.
This directs users to an error page which explains what went wrong and logs it. As well, not-found 404 handler has a two-arity option in which a message can be provided to be logged, allowing the site to hide behind a 404 but still log a descriptive error message in the DB. In summary, use internal-error if we want to tell the user what went wrong, use not-found two-arity if we want to hide from the user what went wrong.